+++ packages/transaction-controller/src/utils/batch.ts
@@ -10,16 +10,22 @@ import {
   isAccountUpgradedToEIP7702,
 } from './eip7702';
 import {
+  getBatchSizeLimit,
   getEIP7702SupportedChains,
   getEIP7702UpgradeContractAddress,
 } from './feature-flags';
 import { validateBatchRequest } from './validation';
-import type {
-  TransactionBatchRequest,
-  TransactionController,
-  TransactionControllerMessenger,
+import {
+  determineTransactionType,
+  type TransactionBatchRequest,
+  type TransactionController,
+  type TransactionControllerMessenger,
 } from '..';
 import { projectLogger } from '../logger';
+import type {
+  NestedTransactionMetadata,
+  TransactionBatchSingleRequest,
+} from '../types';
 import {
   TransactionEnvelopeType,
   type TransactionBatchResult,
@@ -33,6 +39,7 @@ type AddTransactionBatchRequest = {
   getEthQuery: (networkClientId: string) => EthQuery;
   getInternalAccounts: () => Hex[];
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
   request: TransactionBatchRequest;
 };
 
@@ -40,6 +47,7 @@ type IsAtomicBatchSupportedRequest = {
   address: Hex;
   getEthQuery: (chainId: Hex) => EthQuery;
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
 };
 
 const log = createModuleLogger(projectLogger, 'batch');
@@ -58,12 +66,16 @@ export async function addTransactionBatch(
     getChainId,
     getInternalAccounts,
     messenger,
+    publicKeyEIP7702,
     request: userRequest,
   } = request;
+
+  const sizeLimit = getBatchSizeLimit(messenger);
 
   validateBatchRequest({
     internalAccounts: getInternalAccounts(),
     request: userRequest,
+    sizeLimit,
   });
 
   const {
@@ -85,9 +97,14 @@ export async function addTransactionBatch(
     throw rpcErrors.internal('Chain does not support EIP-7702');
   }
 
+  if (!publicKeyEIP7702) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
+
   const { delegationAddress, isSupported } = await isAccountUpgradedToEIP7702(
     from,
     chainId,
+    publicKeyEIP7702,
     messenger,
     ethQuery,
   );
@@ -99,7 +116,12 @@ export async function addTransactionBatch(
     throw rpcErrors.internal('Account upgraded to unsupported contract');
   }
 
-  const nestedTransactions = transactions.map((tx) => tx.params);
+  const nestedTransactions = await Promise.all(
+    transactions.map((tx) =>
+      getNestedTransactionMeta(userRequest, tx, ethQuery),
+    ),
+  );
+
   const batchParams = generateEIP7702BatchTransaction(from, nestedTransactions);
 
   const txParams: TransactionParams = {
@@ -111,6 +133,7 @@ export async function addTransactionBatch(
     const upgradeContractAddress = getEIP7702UpgradeContractAddress(
       chainId,
       messenger,
+      publicKeyEIP7702,
     );
 
     if (!upgradeContractAddress) {
@@ -150,7 +173,16 @@ export async function addTransactionBatch(
 export async function isAtomicBatchSupported(
   request: IsAtomicBatchSupportedRequest,
 ): Promise<Hex[]> {
-  const { address, getEthQuery, messenger } = request;
+  const {
+    address,
+    getEthQuery,
+    messenger,
+    publicKeyEIP7702: publicKey,
+  } = request;
+
+  if (!publicKey) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
 
   const chainIds7702 = getEIP7702SupportedChains(messenger);
   const chainIds: Hex[] = [];
@@ -161,6 +193,7 @@ export async function isAtomicBatchSupported(
     const { isSupported, delegationAddress } = await isAccountUpgradedToEIP7702(
       address,
       chainId,
+      publicKey,
       messenger,
       ethQuery,
     );
@@ -184,4 +217,31 @@ function generateBatchId(): Hex {
   const idString = v4();
   const idBytes = new Uint8Array(parse(idString));
   return bytesToHex(idBytes);
+}
+
+/**
+ * Generate the metadata for a nested transaction.
+ *
+ * @param request - The batch request.
+ * @param singleRequest - The request for a single transaction.
+ * @param ethQuery - The EthQuery instance used to interact with the Ethereum blockchain.
+ * @returns The metadata for the nested transaction.
+ */
+async function getNestedTransactionMeta(
+  request: TransactionBatchRequest,
+  singleRequest: TransactionBatchSingleRequest,
+  ethQuery: EthQuery,
+): Promise<NestedTransactionMetadata> {
+  const { from } = request;
+  const { params } = singleRequest;
+
+  const { type } = await determineTransactionType(
+    { from, ...params },
+    ethQuery,
+  );
+
+  return {
+    ...params,
+    type,
+  };
 }